Added Client main class to open window and show texture
Also added a FloorTile class which other world objects should be based off of. Shaders may be moved out of it in the future since one shader program can be used for all game objects. I know the Client class will probably be merged into the other Main class in the same package but it's currently filled with networking code that I don't understand and I needed somewhere to test my own code