Started MVC integration
New Renderer class which takes in and renders the game model. I've removed all the rest of the render code from the controller (Client) since it should be completely separate. Setup Map and Tile classes in the game in a way that should be slightly easier to use. Generally the two TileSet classes (game and render) should match each other in the order they set up the tiles. I couldn't really think of a better way of doing this while keeping everything separate