Basic next level functionality
Press J to change to next level. Renamed Map to LevelMap since it was interfering with java.util.Map. Also moved players out of Level and into Game since it was better than recreating the player objects every time a level was changed. Game.nextLevel() changes the game state to SCORE when there are no more levels left (I'm sure we'll do something with this in the future)