Refactored the client (the whole client)
Ok this is a big one. The main change is moving most of the screen code out of the Client class into their own respective classes. These get called from the client depending on the current screen being shown. Changing screens is a static method in Client, you pass in the ClientState enum you want to change to along with any arguments you would like to pass to that screen. For these arguments to picked up on the other side the new screen needs to get them from the args parameter which is a variable argument (meaning you can pass in however many arguments you want). You will have to cast the arg[] element to your desired class. I know this isn't exactly type safe but it was the quickest solution I could come up with (feel free to show me a better way). Speaking of static methods; I've made pretty much all the client code for IO/Audio/Window static so that there's no faffing about passing references around and they can be called from anywhere. There was only ever going to one instance of them anyway so it's just easier this way (yes I checked it all still works the same). I may have gone a little static crazy. If you think I have, I don't think it'd be too hard to change them to be singletons instead. Also got rid of the old unused Window, Mouse, and Keyboard classes but did move the window.isKeyPressed methods into Keyboard so that it's a bit clearer. Made some easier methods for getting half the window width and height since that seemed to get used a lot. Player movement is now handled within the gameModel.update method which was kinda necessary for this to work but we did discuss earlier that it might be better this way. Player Ids are also set from the GameModel as we discussed earlier. Removing players is also handled, updating the subsequent player ids. Also I moved the leveleditor package into client since it seemed it should be there.